backend/requirements.txt
# Flask and extensions
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-jwt-extended==4.5.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Google Cloud
google-cloud-storage==2.13.0
google-cloud-aiplatform==1.36.4
google-generativeai==0.3.1

# API and Documentation
flask-restx==1.2.0
marshmallow==3.20.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
pillow==10.1.0
requests==2.31.0

# Development
black==23.11.0
flake8==6.1.0

# 添加Cloud SQL所需的依赖
pymysql==1.1.0
# 如果使用PostgreSQL，则需要
# psycopg2-binary==2.9.9 